create NoopTyperDeduper
