+++ airbyte-integrations/bases/base-typing-deduping/src/main/java/io/airbyte/integrations/base/destination/typing_deduping/NoopTyperDeduper.java
@@ -1,0 +1,18 @@
+package io.airbyte.integrations.base.destination.typing_deduping;
+
+public class NoopTyperDeduper implements TyperDeduper {
+  @Override
+  public void prepareFinalTables() throws Exception {
+
+  }
+
+  @Override
+  public void typeAndDedupe(String originalNamespace, String originalName) throws Exception {
+
+  }
+
+  @Override
+  public void commitFinalTables() throws Exception {
+
+  }
+}
